Fix stack alignment for ukvm target
The x86_64 ABI mandates that (%rsp + 8) be a multiple of 16 on entry to
a user function. GCC relies on this when using SSE instructions and will
ensure correct alignment is preserved in subsequent function calls.

Per testing, virtio target already appears (possibly not intentionally)
to have the correct stack alignment.