Print memory map on boot, add stack guard space
This is useful to be able to quickly eyeball that the memory layout
looks sane and will also aid in debugging addresses in traps / other
crashes.

STACK_GUARD_SIZE is a primitive way of ensuring that (at least) heap
allocation does not overflow into the stack. Ideally we'd want to check
the converse also (stack does not overflow into heap) but the
infrastructure for that is not there yet.