Make no assumptions about type of uint64_t in GUEST_CHECK_PADDR
GUEST_CHECK_PADDR relied on the incorrect assumption that (uint64_t) is
the same type as (unsigned long) and used __builtin_uaddl_overflow()
for checked arithmetic. This does not hold for all compiler
configurations.

This change replaces the checked arithmetic builtin with the generic
__builtin_add_overflow() and adds the corresponding change to the macro
defined for compilers without checked arithmetic builtins.

The documentation for GUEST_CHECK_PADDR is clarified to be explicit
about the expected types, and the format string for the errx() call is
corrected to use PRIx64 / %zu.