ukvm: Improve behaviour when attaching TAP device.
The existing tun_alloc() was inherited from older code, it's purpose is
actually to attach to a pre-existing TAP device. Fix the code to behave
as intended.

User-visible change: If uvkm is run without the specified tap device in
existence then it could create a new one with nothing attached to it.
This would hit an obscure assertion as soon as the guest tried to send a
network packet. With these changes we instead attempt to (*) detect this
case and fail immediately with "No such device".

(*) See the code for the gory details, the Linux TUN/TAP API is a
horrible mess so we rely on undocumented but observed behaviour to try
and make this work.