nolibc migration for tests (fixes #37)
This is  modified copy of the freestanding nolibc.

What has been modified is the removal of the fmt_fp
function of vfprintf.c (with their corresponding flags
e, f, g and ) and the dependency on openlibm (which
means <math.h> is a noop).

The provided test cases tests solo5 malloc.