Standardise license headers and copyright notices
This commit unifies the license headers and copyright notices in Solo5
and ukvm source files. Original source files use one of the following
standard headers:

For the Solo5 unikernel:

    Copyright (c) 2015-2017 Contributors as noted in the AUTHORS file

    This file is part of Solo5, a unikernel base layer.

For ukvm:

    Copyright (c) 2015-2017 Contributors as noted in the AUTHORS file

    This file is part of ukvm, a unikernel monitor.

The header is followed by the standard ISC license text.

Where a source file contains existing third-party licenses or copyright
notices these have been preserved and the Solo5 header has been added to
the top of the file.

Further, this commit updates the copyright notice in the top-level
LICENSE file to read:

    Copyright (c) 2015-2017 Contributors as noted in the AUTHORS file

and adds a new AUTHORS file containing:

1. copyright notices for corporate contributors,
2. a list of individual contributor's names and their affiliation(s),
   if any.

I have populated the AUTHORS file with all contributors to date based on
the information in the git history.