ukvm,kernel: Replace pvclock with simplified tscclock
To enable multi-backend we need a clock that does not depend on any
backend-specific device. This change replaces the use of pvclock by the
ukvm target with:

- a simplified TSC-only implementation of tscclock
- a new hypercall to get wall clock time from the monitor
  (UKVM_HYPERCALL_WALLTIME).
- a new arch-dependent section in struct ukvm_boot_info, used to pass
  the (x86) TSC frequency to the guest.

Note that the monitor will now refuse to run on a host with an unstable
TSC, see the implementation notes in kernel/ukvm/tscclock.c for an
explanation.

Related changes also part of this commit:

- Add a simple test that the clock(s) work(s), in tests/test_time. The
  upper limit on how much is "too much" to sleep may need to be tweaked
  (or the test result ignored) on Travis.
- hypercall_poll() now blocks all signals except SIGINT and SIGTERM
  rather than restarting the ppoll() call which was incorrectly
  assuming (*ts) would be modified by ppoll().