kernel: Factor out arch-specific code
This is the first step for multi-arch support in the ukvm guest.
Arch-specific code (currently only x86_64) is factored out from all over
the codebase into cpu_x86_64.c and cpu_x86_64.h.

The assembly code in cpu.S has been replaced with either inline
functions in cpu_x86_64.h or directly with inline assembly in
cpu_x86_64.c.

As part of this change the virtio early bootstrap (boot.S) is modified
to initialise the CPU into the same state as is used by the ukvm target,
so that the two targets can share the cpu_init() code.