Minimal refresh of Solo5 public APIs
This is a minimal refresh of the Solo5 public APIs and their underlying
backend implementations for ukvm, virtio and muen to fix the most
egregious problems as described in #200.

For a high-level overview of the changes start with the documentation
and public prototypes in kernel/solo5.h.

No functional change intended, hopefully a lot of edge case bugs elided.