kernel: Implement ukvm_do_hypercall for aarch64
AArch64 doesn't support ioport like x86, so we have to use KVM_EXIT_MMIO
to implement hypercall. We have limited the max guest memory size to 4GB
to guarantee the 32-bit pointer used for hypercall is enough. Address space
above 4GB is used for MMIO.

On aarch64, the MMIO address must be 64-bit aligned. Because the MMIO the
memory attributes of MMIO space will be configured to MT_DEVICE_nGnRnE, which
is not allow an unaligned access.