ukvm: Introduce a helper to enable AArch64 guest MMU
If we do not set up the guest MMU then the guest and host will
access memory with a different cacheability which can lead to
data corruption. We need to setup the guest to use the same
cacheability as the host by setting up the MMU within the guest.

In this helper, we configure the guest to use 40-bit VA and IPA
with 4KB page. So we will create a page table to translate VA to
IPA by 4KB granule.