ukvm: Introduce a helper to setup virtual memory mapping for AArch64
We has enabled the guest MMU to make guest to have the same cacheability
as the host. Hence the guest will use virtual address to access memory.
But guest on other architecture are using the physical address. We don’t
want to make guest on aarch64 be special, so we create page tables for
guest to do 1:1 mapping between virtual address and intermediate physical
address.