Travis: Switch to container-based Trusty infrastructure
With the new VM-based CI in place, Travis is more or less just a
fallback. To speed up the Travis builds, switch to using the
Ubuntu Trusty container-based infrastructure.

This loses us sudo access so we can no longer run tests/test_ping_serve,
this is an acceptable compromise for (hopefully) faster builds.