dumpcore: Rewrite to use prstatus_t, optimize
The (host kernel) <-> (debugger) interface and content of NT_PRSTATUS is
codified in <sys/procfs.h>, so rewrite to use it rather than inventing
our own. This also make it clear which parts of dumpcore are
host-os-dependent, leading to easier implementation for other host OSes
and less maintenance in the long run.

While I'm at it, rewrite the actual dumping of the core file to use
mincore() and only dump the pages actually touched by the guest.