Renaming: Directory structure, sources and build products
This is the first step in renaming ukvm to a generic name and the
associated restructuring of Solo5 with introduction of new terminology
for its components. (See #172 for details.)

This commit implements the new directory structure, renaming source
modules and associated build products:

* kernel/X: Moved to bindings/X, now referred to as the "Solo5 bindings
for X". Build products are now named bindings/X/solo5_X.o.
  * kernel/solo5.h: Moved to include/solo5/solo5.h.

* ukvm: Target has been renamed to "hvt". Monitor code is now referred to
as the "hvt tender" and has been moved to tenders/hvt/.
  * ukvm-configure: Now named solo5-hvt-configure.
  * ukvm-bin: Now named solo5-hvt.
  * ukvm/ukvm_guest.h: Renamed to include/solo5/hvt_abi.h, and the
  symlink in tenders/hvt/ has been replaced by a hack in
  solo5-hvt-configure for in-tree builds.

* include-host/: Generated at build time, moved to include/crt/.

No changes are made to any internal or external API names or symbols
used internally at this stage. OPAM package names are likewise unchanged
and OPAM installation is not functional at this stage.

No other functional changes at this stage.