Renaming: OPAM packages
This is the third step in renaming ukvm to a generic name and the
associated restructuring of Solo5 with introduction of new terminology
for its components. (See #172 for details.)

This commit renames the OPAM packages built from the Solo5 repository as
follows:

  solo5-kernel-ukvm: solo5-bindings-hvt
  solo5-kernel-virtio: solo5-bindings-virtio
  solo5-kernel-muen: solo5-bindings-muen

Additional changes:

* The Muen binding gets its own copy of the linker script, this makes it
easier to eyeball that the Muen build runes are correct.
* OPAM installation of header files had a redundant "include/" in the
path, this has been removed.
* include/solo5/hvt_abi.h is also installed to the OPAM solo5-hvt source
directory to allow for correct operation (see solo5-hvt-configure).