Renaming: Internal symbols
This is the second step in renaming ukvm to a generic name and the
associated restructuring of Solo5 with introduction of new terminology
for its components. (See #172 for details.)

The bulk of this commit is a semi-automated replacement of "ukvm_" with
"hvt_" and "UKVM_" with "HVT_", with manual changes to the remaining
cases.

Where the introduction of "hvt_" would cause internal names to read
"hvt_hv_X" the additional "_hv_" has been replaced with just "_".

NOTABLY: "struct ukvm_hv" has been renamed to just "struct hvt" and
"struct ukvm_hvb" (the backend-specific part) has been renamed to
"struct hvt_b". Where "hv" is used as a local variable it has been
renamed to "hvt". The "hvb" local variable case has not been renamed, as
this is not confusing with anything and is only used as "shorthand" for
"hvt->b".

Additionally, this commit covers a few more manual cases where old names
were still in use, and also clarifies the relationship between
bindings/bindings.h and bindings/X/bindings.h in the comments.

Note that the generated VM names used on FreeBSD and OpenBSD have been
changed from "ukvm%d" to "solo5-%d", with %d remaining as the PID of the
solo5-hvt tender. These names are limited to 32 bytes on both platforms
(VM_MAX_NAMELEN, VMM_MAX_NAME_LEN), hence we keep them as short as
possible.

The core file names produced by the dumpcore module have been changed
from "core.ukvm.%d" to "core.solo5-hvt.%d".

No other functional changes.