Escape boot parameters to quoted space-separated format of Parse_argv
Parse_argv.parse is expecting a space separated list of quoted
backslash-escaped strings.  Until now the parameters of argv are just
concatenated with spaces, which results in lost information if argv
elements contain spaces themselves, and the user has to take care of
the quoting and escaping of the boot parameters by themselves.
Together with the escape requirements of typical shells, several
layers of escaping are necessary, which is tedious and error prone.
To avoid this, this change replaces ' ', '"' and '\' with '\ ' '\"' and
'\\' respectively, and joins them with spaces into a single string,
so that Parse_argv.parse can resemble a list of strings identical to
the original argv, which makes additional escaping unnecessary.

Related: #281
Fixes: mirage/mirage-solo5#33