Stack smashing protection
Initialise the __stack_chk_guard value on x86_64 and aarch64 from
crt_init_early(), inlined into _start(). Given that we don't have a
source of randomness, derive a weak pseudo-random value using CPU ticks.
This is a reasonable trade-off for now without diving into the extra
complexity and ABI changes getting randomness from the host would
entail.

__stack_chk_fail() takes care to abort as quickly as possible as the
state of the stack is unknown at this point.

A test_ssp has been added which smashes the stack intentionally, thus
verifying that SSP is working as expected.

<rant>
The ABI that the compiler expects for SSP is *totally* uncodumented, not
standardised anywhere and varies entirely depending on the host OS,
toolchain (GCC/LLVM), and CPU architecture.  This change includes
support only for some of the arch/host/toolchain combinations we care
about, based on reading through both the GCC and LLVM source code and
inspecting code generated by the compiler. Notably, SSP on OpenBSD
toolchains is not supported at this time.
</rant>