virtio,hvt: x86_64: Unmap low memory, disable TLS
This change unmaps low memory (notably at least the zero page) and
disables accidental use of TLS by any code (bindings or application).

On hvt: We unmap the low 64kB, including the guest page tables. The
remainder of the first 1MB (up to X86_GUEST_MIN_BASE) is mapped
read-only, as this is used for early boot input data passed to the guest
from the tender, and the rest is considered "reserved".

For TLS: we add early "C runtime" initialisation code in
bindings/crt_init.h ("called" first thing from _start()) which ensures
that CPU registers related to TLS addressing are zeroed.

Adds the following test cases:

- test_zeropage: verifies receipt of #PF on access to zero page.
- test_notls: verifies that a low TLS slot access results in #PF.

On aarch64 this change only implements the test cases, which have been
verified manually to be correct, and are currently expected to *fail* in
tests.bats. Actually unmapping low memory on aarch64 will be done in a
separate change, at which time these tests cases can be expected to
succeed.