tenders/hvt: Use the similar low memory layout as x86_64 for Arm64
We use the similar low memory layout as x86_64, this will help to enforce
the ABI (with respect to the low 1MB of memory being "reserved"), and
prevent the guest from stomping over it's own page tables.
The difference is that, Arm64 support max 4GB ram currently, so we increase
the PMD table (L2 table) size to 16KB to cover all RAM area.

Signed-off-by: Wei Chen <wei.chen@arm.com>