tenders/hvt: Fix corner case of --mem=1
Asking for 1MB of guest memory resulted in the memory size being rounded
down to zero, which caused an obscure error further on. Check for this
case and adjust to the minimum size.

Additionally, correct an err() call to errx() to squash a confusing
"Success" and replace error messages in the internal
aarch64_setup_memory_mapping() with assertions.