spt: Correct types for sys_...() functions
The sys_...() functions are intentionally weakly typed as they only pass
through values to/from the system call without interpretation. All
integer values are passed as (long) and all pointer values are passed as
(void *). This matches the convention musl libc uses for x86_64 and
aarch64.

This will need to be re-considered for 32-bit archs, and we should also
consider explicitly inlining these functions.