Attempt to work around FreeBSD vmm create/destroy mis-design
More specifically, hw.vmm.destroy needs privileges, so we can no longer
do this in atexit() with dropped privileges.

Instead, look for solo5-YYY before hw.vmm.create, and if it exists,
destroy it (unless it is still running). Testing whether the VM is
running is not possible, we use an exclusive FS lock for this. This
mechanism works under the assumption that solo5-hvt owns the
/dev/vmm/solo5-* namespace.

Contribution was developed pair-programming with @mato at Espresso
Library on Sat 26th January 2019.

Note that this still leaks a VM and its resources on every solo5-hvt
process exit.