bindings: Replace ee_printf() with a better printf()
ee_printf() is unmaintained and does not support standard format
specifiers like %llu which have been introduced in recent commits for
OpenBSD host toolchain portability.

Replace it with the MIT-licensed embedded printf implementation from
https://github.com/mpaland/printf, which is maintained and tested.

The following modifications have been made to mpaland/printf:

1. Remove all public APIs except snprintf() and vsnprintf() as we don't
need them.
2. Disable floating-point support by default.

While we're at it:

- remove the bindings-wide prototype of printf() entirely, everything is
using log() for ages now
- do the "buffering" formerly done in ee_printf() in log(), and tell the
user if the message was truncated

This printf() implementation is about 2kB larger than ee_printf on
x86_64, but that seems like a reasonable price to pay for a better
printf().