Multiple devices support, application manifest
Introduce support for multiple devices and an application manifest.

Abridged list of changes:

The application manifest is generated from a JSON file at unikernel
build time using the newly introduced "mfttool", and linked in to the
binary as an ELF NOTE.

Solo5 network and block APIs are extended to take a solo5_handle_t, and
the solo5_{block,net}_info() calls are replaced with
solo5_{block,net}_acquire(), which are used by the application to
acquire a handle for a device named in the manifest. The solo5_yield()
call has been extended to return a solo5_handle_set_t which is a bitmask
of the (network) handles ready for input.

Tenders (hvt and spt) now require that ALL devices declared in the
application manifest are attached (to a host backing device) at run
time. The --net=IFACE option has been replaced by --net:NAME=IFACE, and
the --disk=FILE option has been replaced by --block:NAME=PATH.

For hvt (tender) and spt (tender/bindings), the ppoll() based polling
mechanism has been replaced with "wait sets" using epoll(7) on Linux and
kqueue(2) on the BSDs. This is necessary to allow for efficient mapping
from file descriptors to solo5_handle_t. Further, on Linux, a timerfd is
used internally in the epoll() wait set to provide accurate timeouts as
per the Solo5 API, since epoll_wait() only accepts timeouts in
milliseconds.

The hvt hypercall ABI has been re-factored and cleaned up to align
better with the Solo5 APIs it implements. The HVT_HYPERCALL_NETINFO and
HVT_HYPERCALL_BLKINFO calls have been removed as their functionality is
now provided by the application manifest passed from tender to guest.

spt and hvt tender mechanisms for registering modules have been unified.