Update virtio to new APIs, application manifest support
This update the virtio bindings to the new Solo5 APIs and uses the
in-built application manifest at run time.

Note that the bindings still support only a single network or block
device internally. Compatibility with the updated APIs is achieved by
allowing only the first valid solo5_{block,net}_acquire() call to
succeed.