Reserve the first manifest entry
In the current design, handles are an index into the manifest, so the
first handle has a value of 0. This is undesirable, since it allows
among other things an unitialized value to be a valid handle.

In order to not use the value 0 as a handle, we mandate that the first
manifest entry (with an index of 0) is reserved and must always be
present. This is done by introducing an entry type of MFT_RESERVED_FIRST,
which is implicitly added to the manifest by mfttool and deliberately
not present in the user-visible JSON definition.