ELF loader: Generic NOTE loading
Make the NOTE loading functionality generic, rather than coupled to
loading just a "MFT1" note. elf_load_mft(...) is now renamed as
elf_load_note(...). The code is also intentionally decoupled from any
dependency on mft_abi.h.

In the process, various NOTE structure alignment issues which were
"papered over" in the initial iteration of this code have been fixed now
that I have a correct understanding of how (and why) the structures are
laid out.

Additional fixes:

- place .note.solo5.manifest in .data (as opposed to .rodata) for virtio
and muen, as these bindings are expected to need to modify the structure
in-place (as opposed to getting a copy from a tender).
- improve virtio built-in manifest initialisation, and introduce a
global variable for access to the validated struct mft.
- document the ELF functions some more.