Ensure that guest-executable pages are not host-executable.
During discussion of Linux/KVM and mprotect() interaction, it was
pointed out to me that guest-executable pages should not also be marked
PROT_EXEC on the host side. Fix this using the newly introduced
t_guest_mprotect() infrastructure. At the same time, clarify the exact
behaviour on the various backends (hvt, spt) in comments.