Add a "make distrib" for building release tarballs
Warning, Makefile and Git hackery here.

Add a "make distrib", to be used for building release tarballs.
Generates a static "solo5_version.h.distrib" which is included in the
release tarball. This is then used by scripts/gen_version.h.sh as the
source of truth for generating "solo5_version.h", so that builds from a
release tarball work without Git, but contain meaningful version
information.