Implement tender/bindings ABI versioning
This change introduces an additional Solo5-owned ELF note, "ABI1", which
is used to embed information about the backend target (hvt, spt, ...)
and internal tender/bindings ABI version of a unikernel binary linked
with the Solo5 bindings.

The internal ABI version is defined as a single uint32_t value, starting
at 1. The spt and hvt tenders now refuse to load a binary whose ABI
version does not match that of the tender.