Add test_mft_maxdevices, "require" libseccomp >= 2.3.3
Add a test case with the maximum allowed number of manifest devices.

This test has uncovered a bug in (at least) libseccomp older than 2.3.3
where libseccomp would wrongly issue a brk() call (presumably from
malloc) *after* the seccomp filter is installed yet during the execution
of seccomp_load(), aborting the spt tender with SIGSYS. Add a warning
and don't build spt if libseccomp >= 2.3.3 is not available on the host.

Additionally, some minor refactoring in tests.bats, to not create
backing storage for block devices unless the test requires it.