Add diagnostic output to test_time
test_time has been failing intermittently on Debian 10 VM-based CI with
"slept too little". This may be due to scheduling latency/load of the
virtualized environment, or it may be a more subtle issue. Add some
diagnostic printf()s to the test.

This also opens the door to use of printf() in tests generally, which is
useful.