mft: Enforce const-ness
The built-in manifest may be in a read-only section (e.g. for muen).
In order to enforce that, we should deal with const-ness as well as we
can within the constraints of C:

- mft_validate() was previously modifying the manifest in-place. Don't
do that, and take a const pointer as input.
- mft_get_by_...() now take a const pointer, but return a non-const
since otherwise we'd have to provide two versions of these functions.
- mft_get_builtin_mft1() now has an explicit _unconst version, and uses
const pointers throughout as appropriate.
- Muen and virtio bindings have been modified to suit.