bindings: Add some more const-ness for mft, spt/hvt ABI
No functional changes.

Inspired by #404, add some more const-ness to (struct mft *) and (struct
mft_entry *) used by the bindings.

For hvt and spt, also make the argument passed to _start() and its
pointer members const, as it is in low memory and will result in a fault
if accidentally modified. The rest is just the ripple-down effect of
these changes.