Stop host kernels from executing Solo5 binaries
This is implemented for hvt, spt, muen and virtio bindings by:

- Adding a PT_INTERP and .interp in binding-specific linker scripts.
- Populating .interp via DECLARE_ELF_INTERP in binding-specific start.c.
- The content of PT_INTERP is "/nonexistent/solo5/". See comments in
elf_abi.h explaining the choice.

The Genode bindings have not been changed, as they seem to do their own
thing regarding PT_INTERP.

With this change if a Solo5 unikernel binary is run directly (e.g.
./test_hello.hvt) on a Linux or FreeBSD host, the host kernel will
not attempt to execute the binary and will fail with a more or less
sensible error message rather than produce a segmentation fault or
similar.

OpenBSD hosts will fail to load the ELF executable, but then pass it on
to /bin/sh, there is nothing we can do about this behaviour.

This provides both for a nicer "naive user" experience, and is a
security improvement against accidentally exposing the host kernel to
unikernel binaries from untrusted sources.