configure.sh: Improve CRT header copying on BSDs
The *BSD C runtime header copying code is somewhat fragile and requires
manually keeping track of dependencies of each header that gets copied.

This change uses cc -M to gather a list of dependencies of all the
requested header files, and then cpio to copy the result including
intermediate directories to the target directory.

While I'm at it, remove osreldate.h from SRCS= on FreeBSD, since
as far as I can tell there's nothing that depends on that.