hvt: Fix hang in HVT_HYPERCALL_POLL
Calling solo5_yield(deadline, ...) with a deadline that has passed
results in HVT_HYPERCALL_POLL passing a struct itimerspec with an
it_value of zero in both fields to timerfd_settime(), which disarms the
timer, which causes the subsequent call to epoll_pwait() to hang forever
if no other file descriptors (external events) have been registered with
the waitset.

Fix this case cheaply by foregoing 1 ns of precision and always setting
the LSB of it_value.tv_nsec to 1.