Xen bindings: Deliver event channel upcall via APIC vector
As discussed in [1], the preferred way of registering event channel
upcalls with Xen for modern PVH guests is via
HVMOP_set_evtchn_upcall_vector which delivers the interrupt via an APIC
vector.

Therefore, we replace the legacy PIC initialisation code with minimal
APIC initialisation code, and switch to the modern registration method.

[1] https://lists.xenproject.org/archives/html/mirageos-devel/2020-06/msg00010.html