Xen: time: Fix potential races in solo5_yield()
Based on further discussion on the Xen mailing list, and review of the
equivalent code in Mirage/Xen, fix solo5_yield() to wait for delivery of
a timer event in a race-free way.

For the avoidance of confusion, note that this code is not used by
Mirage/Xen, but nevertheless must be correct for the limited scope of
solo5_yield() in the Xen bindings so that tests pass reliably.