spt: Load seccomp filter manually
libseccomp's seccomp_load() calls free(), which, depending on the libc
malloc() behaviour and generated BPF filter size may call brk(), which
we do not have in our seccomp filter. This leads to a crash during
solo5-spt start-up on some systems.

Work around this by exporting the generated BPF and loading it into the
kernel manually with seccomp().