virtio: add missing reset for net and blk devices
The configuration functions for virtio blk and net were missing the first step
of the device initialization sequence (virtio spec 1.0).

	3.1.1 Driver Requirements: Device Initialization
	The driver MUST follow this sequence to initialize a device:
	1. reset the device.
	2. ... <=== we seem to follow everything nicely after this

The consequence of this is that a vmm (like QEMU) couldn't boot using a disk
image exposed as a virtio-blk device. This was reported in issue 483.  The
issue is that the boot loader (before solo5 discover the virtio devices)
initializes the device, negotiates features, and reads the kernel from it. The
solo5 kernel then tries to negotiate features again, which is fine, except that
it does it without a reset. This is wrong according to the spec:

	2.2 Feature Bits
	... The only way to renegotiate is to reset the device.

Tested:
	cd tests
	../scripts/virtio-mkimage/solo5-virtio-mkimage.sh -f raw \
		disk-net.img test_net/test_net.virtio
	sudo qemu-system-x86_64 -drive file=disk-net.img,if=virtio,format=raw \
		-cpu Westmere -m 128 -nodefaults -no-acpi -display none \
		-serial stdio -device virtio-net,netdev=n0 \
		-netdev tap,id=n0,ifname=tap100,script=no,downscript=no \
		-device isa-debug-exit
	ping 10.0.0.2 # works