Remove Genode bindings
The Genode bindings have been unmaintained for some time, and are
incompatible with the upcoming LLVM/Clang based toolchain. Remove them
for now, they can be reinstated in the future if there is interest.

Note that GENODE_ABI_TARGET is not removed, since once allocated the
value should not be re-used by any future bindings.