Delay checking capacity until block_size is set
We can't check in `block_attach()` because we can't be sure what the
block size is until we have handled all command line arguments. Instead,
we do the capacity checks against the block size in `setup()`.  As the
file path is no longer available we report errors using the name from
the manifest instead of the file path.